templates/ai-tool-app/1.0.0/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
AI工具应用主程序
"""

import yaml
import requests
from typing import Dict, List, Any, Optional


class AIToolApp:
    """AI工具应用类"""
    
    def __init__(self, config_path: str = "config.yaml"):
        """初始化AI工具应用"""
        self.config = self.load_config(config_path)
    
    def load_config(self, config_path: str) -> Dict:
        """加载配置文件"""
        try:
            with open(config_path, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        except Exception as e:
            raise Exception(f"配置文件加载失败: {e}")
    
    def process(self, input_text: str, options: Optional[Dict] = None) -> Dict[str, Any]:
        """处理AI任务"""
        try:
            options = options or {}
            
            # TODO: 实现具体的AI处理逻辑
            result = self._perform_ai_task(input_text, options)
            
            return {
                "status": "success",
                "output": result["output"],
                "metadata": result["metadata"]
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e)
            }
    
    def _perform_ai_task(self, input_text: str, options: Dict) -> Dict[str, Any]:
        """执行实际AI任务"""
        ai_config = self.config.get('ai_service', {})
        provider = ai_config.get('provider', 'openai')
        
        # TODO: 根据配置的AI服务提供商实现处理
        if provider == 'openai':
            return self._process_with_openai(input_text, options)
        elif provider == 'azure':
            return self._process_with_azure(input_text, options)
        elif provider == 'local':
            return self._process_with_local_model(input_text, options)
        else:
            raise Exception(f"不支持的AI服务提供商: {provider}")
    
    def _process_with_openai(self, input_text: str, options: Dict) -> Dict[str, Any]:
        """使用OpenAI处理"""
        # TODO: 实现OpenAI API调用
        # 这里是示例代码
        
        processed_text = f"AI处理结果: {input_text}"
        
        return {
            "output": processed_text,
            "metadata": {
                "tokens_used": len(input_text) + len(processed_text),
                "processing_time": "0.5s",
                "model": self.config.get('ai_service', {}).get('model', 'gpt-3.5-turbo')
            }
        }
    
    def _process_with_azure(self, input_text: str, options: Dict) -> Dict[str, Any]:
        """使用Azure OpenAI处理"""
        # TODO: 实现Azure OpenAI API调用
        processed_text = f"Azure AI处理结果: {input_text}"
        
        return {
            "output": processed_text,
            "metadata": {
                "tokens_used": len(input_text) + len(processed_text),
                "processing_time": "0.6s",
                "model": "azure-gpt"
            }
        }
    
    def _process_with_local_model(self, input_text: str, options: Dict) -> Dict[str, Any]:
        """使用本地模型处理"""
        # TODO: 实现本地模型调用
        processed_text = f"本地模型处理结果: {input_text}"
        
        return {
            "output": processed_text,
            "metadata": {
                "tokens_used": len(input_text) + len(processed_text),
                "processing_time": "1.0s",
                "model": "local-model"
            }
        }


def main():
    """主函数"""
    app = AIToolApp()
    
    # 示例AI处理
    result = app.process("帮我分析这个文本的情感", {"temperature": 0.7})
    print(result)


if __name__ == "__main__":
    main()