templates/ai-tool-app/requirements.txt
pyyaml>=6.0
requests>=2.28.0
openai>=1.0.0
transformers>=4.21.0
torch>=2.0.0
numpy>=1.24.0